Refactor domain entities and interfaces
Refactored KeyEntity classes into separate files and updated interfaces accordingly. Simplified Entity class by removing unnecessary derived classes and introduced IEquatable for more robust type comparison. Added new methods to EntityTypeBuilderExtensions for handling default timestamps.
