+++ Domain/src/Extensions/EntityTypeBuilderExtensions.cs
@@ -1,7 +1,5 @@
 ﻿// Copyright (c) 2014-2024 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-using System.Numerics;
-
 using Wangkanai.Domain;
 
 namespace Microsoft.EntityFrameworkCore.Metadata.Builders;
@@ -13,6 +11,12 @@ public static class EntityTypeBuilderExtensions
 		=> builder.Property(x => x.Id)
 		          .ValueGeneratedOnAdd();
 
+	// public static void NewKeyOnAdd<T, TKey>(this EntityTypeBuilder<T> builder)
+	// 	where T : class, IEntity<TKey>
+	// 	where TKey : IComparable<TKey>, IEquatable<TKey>
+	// 	=> builder.Property(x => x.Id)
+	// 	          .ValueGeneratedOnAdd();
+
 	public static void NewKeyOnAdd<T>(this EntityTypeBuilder<T> builder)
 		where T : class, IEntity<int>
 		=> builder.Property(x => x.Id)
@@ -23,6 +27,12 @@ public static class EntityTypeBuilderExtensions
 		=> builder.Property(x => x.Created)
 		          .HasDefaultValue(DateTime.Now)
 		          .ValueGeneratedOnAdd();
+
+	public static void HasDefaultUpdated<T>(this EntityTypeBuilder<T> builder)
+		where T : class, IUpdatedEntity
+		=> builder.Property(x => x.Updated)
+		          .HasDefaultValue(DateTime.Now)
+		          .ValueGeneratedOnUpdate();
 
 	public static void HasDefaultCreatedAndUpdated<T>(this EntityTypeBuilder<T> builder)
 		where T : class, ICreatedEntity, IUpdatedEntity
