+++ Domain/src/Interfaces/IKeyGuidEntity.cs
@@ -2,12 +2,4 @@
 
 namespace Wangkanai.Domain;
 
-public interface IKeyByteEntity : IEntity<byte>;
-
-public interface IKeyIntEntity : IEntity<int>;
-
-public interface IKeyLongEntity : IEntity<long>;
-
 public interface IKeyGuidEntity : IEntity<Guid>;
-
-public interface IKeyStringEntity : IEntity<string>;
