+++ Domain/src/Interfaces/IUnitOfWork.cs
@@ -1,7 +1,5 @@
 // Copyright (c) 2014-2023 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-using System.Threading.Tasks;
-
 namespace Wangkanai.Domain;
 
 public interface IUnitOfWork
@@ -9,7 +7,7 @@ public interface IUnitOfWork
 	int Commit();
 }
 
-public interface IAsyncUnitOfWork
+public interface IUnitOfWorkAsync
 {
 	Task<int> CommitAsync();
 }
